requirements.txt
# Google Trends API
pytrends>=4.9.2

# Google Sheets API
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Variables de entorno
python-dotenv>=1.0.0

# Utilidades
pandas>=2.0.0

# Base de datos Turso (SQLite cloud)
libsql>=0.1.11

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
